remove config & update args logic
