+++ device_doctor/lib/src/health.dart
@@ -1,0 +1,28 @@
+// Copyright 2020 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// Result of a health check for a specific parameter.
+class HealthCheckResult {
+  HealthCheckResult.success([this.details]) : succeeded = true;
+  HealthCheckResult.failure(this.details) : succeeded = false;
+  HealthCheckResult.error(dynamic error, dynamic stackTrace)
+      : succeeded = false,
+        details = 'ERROR: $error\n${stackTrace ?? ''}';
+
+  final bool succeeded;
+  final String details;
+
+  @override
+  String toString() {
+    StringBuffer buf = StringBuffer(succeeded ? 'succeeded' : 'failed');
+    if (details != null && details.trim().isNotEmpty) {
+      buf.writeln();
+      // Indent details by 4 spaces
+      for (String line in details.trim().split('\n')) {
+        buf.writeln('    $line');
+      }
+    }
+    return '$buf';
+  }
+}
