+++ device_doctor/lib/src/utils.dart
@@ -18,28 +18,3 @@ class BuildFailedError extends Error {
   @override
   String toString() => message;
 }
-
-/// Result of a health check for a specific parameter.
-class HealthCheckResult {
-  HealthCheckResult.success([this.details]) : succeeded = true;
-  HealthCheckResult.failure(this.details) : succeeded = false;
-  HealthCheckResult.error(dynamic error, dynamic stackTrace)
-      : succeeded = false,
-        details = 'ERROR: $error\n${stackTrace ?? ''}';
-
-  final bool succeeded;
-  final String details;
-
-  @override
-  String toString() {
-    StringBuffer buf = StringBuffer(succeeded ? 'succeeded' : 'failed');
-    if (details != null && details.trim().isNotEmpty) {
-      buf.writeln();
-      // Indent details by 4 spaces
-      for (String line in details.trim().split('\n')) {
-        buf.writeln('    $line');
-      }
-    }
-    return '$buf';
-  }
-}
