+++ device_doctor/test/src/device_test.dart
@@ -5,7 +5,7 @@
 import 'package:test/test.dart';
 
 import 'package:device_doctor/src/device.dart';
-import 'package:device_doctor/src/utils.dart';
+import 'package:device_doctor/src/health.dart';
 
 import 'fake_ios_device.dart';
